# Performance: inline them all!!1!