Seems hokey to have to tell the compiler to inline accessors, and obviously leads to a larger build output (`.dll` is roughly 2x the size  or so), but it seems to make for some ok gains given how often the accessors are used throughout the library?

* before AggressiveInlining

|                                            Method |      Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------------------------------- |----------:|---------:|---------:|------:|------:|------:|----------:|
|         pocketken.H3.H3Index.RotateClockwise(hex) | 147.38 ns | 1.142 ns | 1.068 ns |     - |     - |     - |         - |
|  pocketken.H3.H3Index.RotateCounterClockwise(hex) | 138.16 ns | 0.681 ns | 0.604 ns |     - |     - |     - |         - |
|        pocketken.H3.H3Index.RotateClockwise(pent) |  74.40 ns | 0.865 ns | 0.767 ns |     - |     - |     - |         - |
| pocketken.H3.H3Index.RotateCounterClockwise(pent) |  70.87 ns | 0.608 ns | 0.569 ns |     - |     - |     - |         - |
| &#39;pocketken.H3.GetKRingFast(hex, k = 50)&#39; |   593.2 us |  4.04 us |  3.78 us |  66.4063 | 33.2031 |       - |  547.92 KB |
| &#39;pocketken.H3.GetKRingSlow(hex, k = 50)&#39; | 5,846.9 us | 26.14 us | 24.45 us | 179.6875 | 85.9375 | 85.9375 | 1634.09 KB |
|  pocketken.H3.Compact |  356.1 ms | 6.75 ms | 6.63 ms | 11000.0000 | 3000.0000 |         - | 243.51 MB |


* after AggressiveInlining

|                                            Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------------------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|
|         pocketken.H3.H3Index.RotateClockwise(hex) | 58.46 ns | 0.373 ns | 0.311 ns |     - |     - |     - |         - |
|  pocketken.H3.H3Index.RotateCounterClockwise(hex) | 64.83 ns | 0.737 ns | 0.690 ns |     - |     - |     - |         - |
|        pocketken.H3.H3Index.RotateClockwise(pent) | 28.37 ns | 0.027 ns | 0.022 ns |     - |     - |     - |         - |
| pocketken.H3.H3Index.RotateCounterClockwise(pent) | 28.38 ns | 0.306 ns | 0.271 ns |     - |     - |     - |         - |
|  &#39;pocketken.H3.GetKRingFast(hex, k = 50)&#39; |   438.1 μs |  4.23 μs |  3.75 μs |  66.4063 | 33.2031 |       - |  547.92 KB |
|  &#39;pocketken.H3.GetKRingSlow(hex, k = 50)&#39; | 3,820.5 μs | 23.49 μs | 20.83 μs | 179.6875 | 89.8438 | 89.8438 | 1634.09 KB |
| pocketken.H3.Compact | 333.2 ms | 3.96 ms | 3.70 ms | 11000.0000 | 3000.0000 |         - | 243.51 MB |

Will put them in place for now until someone smarter comes along and can provide some better guidance on what we should/shouldn't be doing here...  Definitely feels a bit excessive and dirty to have to do this.